+++ CHANGELOG.md
@@ -3,8 +3,14 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+## 0.4.1 (2020-11-20)
+
 - Making ref patch removes panel a split panel instead of a v-box
 - Fix ref states patches and squash functionality
+
+## 0.4.0 (2020-11-17)
+
+- Implement ref tracking
 
 ## 0.3.7 (2020-11-12)